Make send_error() a bot method instead of commands cog method
